Refactor AddRecipe component structure and styles
- Moved AddRecipe, AddIngredients, and AddInstructions components to a new directory structure under `src/components/AddRecipe/`.
- Deleted old CSS and JSX files for AddRecipe, AddIngredients, and AddInstructions.
- Created new CSS and JSX files for AddRecipe, AddIngredients, and AddInstructions with updated styles and structure.
- Updated Navbar and RecipeCard components to reflect new paths and styles.
- Removed unused CSS files for Navbar and RecipeCard.
- Enhanced Home component layout and styles for better user experience.
