+++ src/components/Home/Home.css
@@ -1,0 +1,47 @@
+.home {
+  background: rgba(255, 248, 240, 0.9); 
+  border-radius: 16px;
+  padding: 24px;
+  max-width: 800px;
+  font-family: system-ui, -apple-system, Segoe UI, Roboto, Helvetica, Arial, sans-serif;
+  color: #111827;
+}
+
+.home__header {
+  background: rgba(229, 231, 235, 0.8);
+  border-radius: 12px;
+}
+
+.btn {
+  display: inline-flex;
+  padding: 10px 14px;
+  border-radius: 10px;
+  border: 1px solid transparent;
+}
+
+.btn:active { transform: translateY(1px); }
+
+.btn-primary {
+  background: #111827;
+  color: #fff;
+  box-shadow: 0 6px 140px rgba(17, 24, 39, 0.18);
+  margin-top: 20px;
+}
+
+.empty-home- {
+  margin-top: 40px;
+  padding: 40px;
+  border: 1px dashed #e5e7eb;
+  border-radius: 14px;
+  text-align: center;
+  background: #fafafa;
+}
+.empty-home-message { color: #6b7280; margin: 6px 0 16px; }
+
+.card-grid {
+  display: grid;
+  gap: 18px;
+  grid-template-columns: repeat(3, 1fr);
+}
+
+
