+++ src/components/RecipeCard/RecipeCard.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import "./RecipeCard.css";
+
+const RecipeCard = ({ name, time }) => {
+  return (
+    <div className="recipe-card">
+      <h3>{name}</h3>
+      <p>{time} mins</p>
+    </div>
+  );
+};
+
+export default RecipeCard;
